logging started

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { Task } from './task.entity';
 import { InjectRepository } from '@nestjs/typeorm/dist';
 import { TaskRepository } from './task.repository';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -17,8 +18,11 @@ export class TasksService {
   //   return this.tasks;
   // }
 
-  getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  getTasks(
+    filterDto: GetTaskFilterDto,
+    user: User,
+  ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
   // getTaskById(id: string): Task {
@@ -29,8 +33,14 @@ export class TasksService {
   //   return found;
   // }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User,
+  ): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: {
+      id,
+      userId: user.id,
+     }});
 
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -39,20 +49,26 @@ export class TasksService {
     return found;
   }
 
-  createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-      return this.taskRepository.createTask(createTaskDto);
+  createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+      return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result =  await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result =  await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     task.status = status;
     await task.save();
     return task;
